api/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
pydantic[email]==2.7.1
pydantic-settings==2.2.1
PyJWT==2.8.0
bcrypt==4.1.3
python-multipart==0.0.9
email-validator==2.1.1
